refactor: mock table columns
